+++ src/app/admin/components/certifiedProduct/certifiedProduct/edit.controller.js
@@ -132,26 +132,28 @@
             vm.cp.sedTestingEnd = vm.cp.sedTestingEndDate.getTime();
             if (vm.workType === 'manage') {
                 vm.isSaving = true;
-                commonService.updateCP(vm.cp)
-                    .then(function (response) {
-                        if (!response.status || response.status === 200) {
-                            $uibModalInstance.close(response);
-                        } else {
-                            vm.errors = [response.error];
-                            vm.isSaving = false;
-                        }
-                    },function (error) {
-                        vm.errors = [];
-                        if (error.data) {
-                            if (error.data.error && error.data.error.length > 0)
-                                vm.errors.push(error.data.error);
-                            if (error.data.errorMessages && error.data.errorMessages.length > 0)
-                                vm.errors = vm.errors.concat(error.data.errorMessages);
-                            if (error.data.warningMessages && error.data.warningMessages.length > 0)
-                                vm.errors = vm.errors.concat(error.data.warningMessages);
-                        }
+                commonService.updateCP({
+                    listing: vm.cp,
+                    banDeveloper: vm.banDeveloper
+                }).then(function (response) {
+                    if (!response.status || response.status === 200) {
+                        $uibModalInstance.close(response);
+                    } else {
+                        vm.errors = [response.error];
                         vm.isSaving = false;
-                    });
+                    }
+                },function (error) {
+                    vm.errors = [];
+                    if (error.data) {
+                        if (error.data.error && error.data.error.length > 0)
+                            vm.errors.push(error.data.error);
+                        if (error.data.errorMessages && error.data.errorMessages.length > 0)
+                            vm.errors = vm.errors.concat(error.data.errorMessages);
+                        if (error.data.warningMessages && error.data.warningMessages.length > 0)
+                            vm.errors = vm.errors.concat(error.data.warningMessages);
+                    }
+                    vm.isSaving = false;
+                });
             } else if (vm.workType === 'confirm') {
                 $uibModalInstance.close(vm.cp);
             }