+++ src/app/search/search.controller.js
@@ -5,7 +5,7 @@
         .controller('SearchController', SearchController);
 
     /** @ngInject */
-    function SearchController ($analytics, $filter, $localStorage, $location, $log, $rootScope, $scope, $timeout, $uibModal, cfpLoadingBar, commonService, utilService, CACHE_TIMEOUT, RELOAD_TIMEOUT) {
+    function SearchController ($analytics, $filter, $localStorage, $location, $log, $rootScope, $scope, $timeout, $uibModal, commonService, utilService, CACHE_TIMEOUT, RELOAD_TIMEOUT) {
         var vm = this;
 
         vm.browseAll = browseAll;
@@ -17,7 +17,6 @@
         vm.loadResults = loadResults;
         vm.registerAllowAll = registerAllowAll;
         vm.registerClearFilter = registerClearFilter;
-        vm.registerClearTerm = registerClearTerm;
         vm.registerRestoreComponents = registerRestoreComponents;
         vm.registerRestoreState = registerRestoreState;
         vm.registerSearch = registerSearch;
@@ -25,7 +24,6 @@
         vm.statusFont = utilService.statusFont;
         vm.triggerAllowAll = triggerAllowAll;
         vm.triggerClearFilters = triggerClearFilters;
-        vm.triggerClearTerm = triggerClearTerm;
         vm.triggerRestoreState = triggerRestoreState;
         vm.triggerSearch = triggerSearch;
         vm.viewCertificationStatusLegend = viewCertificationStatusLegend;
@@ -54,7 +52,6 @@
             vm.restoreStateHs = [];
             vm.isLoading = true;
             vm.isPreLoading = true;
-            cfpLoadingBar.start();
 
             manageStorage();
             populateSearchOptions();
@@ -88,18 +85,19 @@
                 'Terminated by ONC': false
             }
         };
+        vm.retired = {
+            acb: {'CCHIT': true, 'SLI Global': true, 'Surescripts LLC': true}
+        };
 
         function browseAll () {
-            $analytics.eventTrack('Browse All');
+            $analytics.eventTrack('Browse All', { category: 'Search' });
             vm.triggerClearFilters();
-            vm.triggerClearTerm();
             vm.activeSearch = true;
             setTimestamp();
         }
 
         function clear () {
             vm.triggerClearFilters();
-            vm.triggerClearTerm();
             vm.activeSearch = false;
             if (vm.searchForm) {
                 vm.searchForm.$setPristine();
@@ -136,9 +134,6 @@
 
         function loadResults() {
             commonService.getAll().then(function (response) {
-                if (vm.isPreLoading) {
-                    cfpLoadingBar.start();
-                }
                 var results = response.results;
                 for (var i = 0; i < results.length; i++) {
                     results[i].mainSearch = [results[i].developer, results[i].product, results[i].acbCertificationId, results[i].chplProductNumber].join('|');
@@ -175,16 +170,6 @@
             vm.clearFilterHs.push(handler);
             var removeHandler = function () {
                 vm.clearFilterHs = vm.clearFilterHs.filter(function (aHandler) {
-                    return aHandler !== handler;
-                });
-            };
-            return removeHandler;
-        }
-
-        function registerClearTerm (handler) {
-            vm.clearTerm = [handler]
-            var removeHandler = function () {
-                vm.clearTerm = vm.clearTerm.filter(function (aHandler) {
                     return aHandler !== handler;
                 });
             };
@@ -249,13 +234,6 @@
             vm.triggerSearch();
         }
 
-        function triggerClearTerm () {
-            angular.forEach(vm.clearTerm, function (handler) {
-                handler();
-            });
-            vm.triggerSearch();
-        }
-
         function triggerRestoreState () {
             if ($localStorage.searchTableState) {
                 var state = angular.fromJson($localStorage.searchTableState);
@@ -289,8 +267,8 @@
         }
 
         function viewPreviouslyCompared (doNotSearch) {
-            $analytics.eventTrack('View Previously Compared');
             if (!doNotSearch) {
+                vm.triggerClearFilters();
                 vm.triggerAllowAll();
             }
             $localStorage.viewingPreviouslyCompared = true;
@@ -300,13 +278,14 @@
                 vm.previouslyIds.push({value: id, selected: true})
             });
             if (!doNotSearch) {
+                $analytics.eventTrack('View Previously Compared', { category: 'Search' });
                 vm.triggerSearch();
             }
         }
 
         function viewPreviouslyViewed (doNotSearch) {
-            $analytics.eventTrack('View Previously Viewed');
             if (!doNotSearch) {
+                vm.triggerClearFilters();
                 vm.triggerAllowAll();
             }
             $localStorage.viewingPreviouslyViewed = true;
@@ -316,6 +295,7 @@
                 vm.previouslyIds.push({value: id, selected: true})
             });
             if (!doNotSearch) {
+                $analytics.eventTrack('View Previously Viewed', { category: 'Search' });
                 vm.triggerSearch();
             }
         }
@@ -375,17 +355,19 @@
 
         function populateSearchOptions () {
             vm.lookaheadSource = {all: [], developers: [], products: []};
-            commonService.getSearchOptions()
+            commonService.getSearchOptions(true)
                 .then(function (options) {
-                    if (vm.isPreLoading) {
-                        cfpLoadingBar.start();
-                    }
-
                     vm.searchOptions = options;
                     var i;
                     options.practiceTypes = [];
                     for (i = 0; i < options.practiceTypeNames.length; i++) {
                         options.practiceTypes.push(options.practiceTypeNames[i].name);
+                    }
+                    for (i = 0; i < options.certBodyNames.length; i++) {
+                        if (options.certBodyNames[i].name === 'Pending') {
+                            options.certBodyNames.splice(i,1);
+                            break;
+                        }
                     }
                     for (i = 0; i < options.certificationStatuses.length; i++) {
                         if (options.certificationStatuses[i].name === 'Pending') {
@@ -415,11 +397,9 @@
                 if (difference > CACHE_TIMEOUT) {
                     vm.activeSearch = false;
                 } else {
-                    cfpLoadingBar.start();
                     $timeout(
                         function () {
                             vm.triggerRestoreState();
-                            cfpLoadingBar.complete();
                         },
                         RELOAD_TIMEOUT
                     );
@@ -432,7 +412,7 @@
         }
 
         function setFilterInfo (refineModel) {
-            var i;
+            var i, obj;
             vm.refineModel = angular.copy(refineModel);
             vm.filterItems = {
                 pageSize: '50',
@@ -444,7 +424,14 @@
             };
             vm.searchOptions.certBodyNames = $filter('orderBy')(vm.searchOptions.certBodyNames, 'name');
             for (i = 0; i < vm.searchOptions.certBodyNames.length; i++) {
-                vm.filterItems.acbItems.push({value: vm.searchOptions.certBodyNames[i].name, selected: vm.defaultRefineModel.acb[vm.searchOptions.certBodyNames[i].name]});
+                obj = {
+                    value: vm.searchOptions.certBodyNames[i].name,
+                    selected: vm.defaultRefineModel.acb[vm.searchOptions.certBodyNames[i].name]
+                };
+                if (vm.retired.acb[vm.searchOptions.certBodyNames[i].name]) {
+                    obj.display = obj.value + ' (Retired)';
+                }
+                vm.filterItems.acbItems.push(obj);
             }
             vm.searchOptions.editions = $filter('orderBy')(vm.searchOptions.editions, 'name');
             for (i = 0; i < vm.searchOptions.editions.length; i++) {