+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@ _Date TBD_
 * Allow optionally banning Developers on appropriate status change
 * Tweaked GTM events and expanded the events that are tracked
 * Show description instead of just name on Functionality Tested & Standard
+* Changed description of "Non-conformities" file on download page
 
 ### Bugs fixed
 * Fixed bug related to Merged Developer status history