+++ RELEASE_NOTES.md
@@ -8,10 +8,27 @@
 ## Version TBD
 _Date TBD_
 
+### New Features
+* Allow optionally banning Developers on appropriate status change
+* Tweaked GTM events and expanded the events that are tracked
+* Show description instead of just name on Functionality Tested & Standard
+* Changed description of "Non-conformities" file on download page
+
+### Bugs fixed
+* Fixed bug related to Merged Developer status history
+* Fixed bug related to too long Name/Additional Software in CMS PDF generation
+* Viewing previously viewed/compared clears all filters to allow it
+
+---
+
+## Version 8.2.0
+_24 April 2017_
+
 ### New features
 * Added Google Tag Manager event tracking
 * Added POC to Product level; refactored Contact as separate component
 * Allow splitting of Products along Version lines
+* Show retired/deleted ONC-ACBs on main page search filter
 
 ---